Initial 1.2 work + minor fixes
