+++ package.json
@@ -18,24 +18,30 @@
   },
   "dependencies": {
     "protobufjs": "^6.10.2",
-    "rxjs": "^6.6.3"
+    "rxjs": "^6.6.6"
   },
   "devDependencies": {
+    "@types/jest": "^26.0.20",
     "@types/web-bluetooth": "0.0.9",
+    "jest": "^26.6.3",
     "pre-commit": "^1.2.2",
     "prettier": "^2.2.1",
-    "typedoc": "^0.20.25",
-    "typescript": "^4.1.5"
+    "ts-jest": "^26.5.2",
+    "typedoc": "^0.20.28",
+    "typescript": "^4.2.2"
   },
   "scripts": {
     "build": "tsc",
     "generate-docs": "typedoc src/index.ts",
     "lint": "prettier --check .",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "pre-commit": [
     "lint"
   ],
+  "jest": {
+    "preset": "ts-jest"
+  },
   "keywords": [
     "meshtastic bluetooth http library meshtasticjs"
   ],
