+++ src/protobufs/portnums.ts
@@ -1,0 +1,18 @@
+export enum PortNumEnum {
+  UNKNOWN_APP = 0,
+  TEXT_MESSAGE_APP = 1,
+  REMOTE_HARDWARE_APP = 2,
+  POSITION_APP = 3,
+  NODEINFO_APP = 4,
+  ROUTING_APP = 5,
+  ADMIN_APP = 6,
+  REPLY_APP = 32,
+  IP_TUNNEL_APP = 33,
+  ENVIRONMENTAL_MEASUREMENT_APP = 34,
+  SERIAL_APP = 64,
+  STORE_FORWARD_APP = 65,
+  RANGE_TEST_APP = 66,
+  PRIVATE_APP = 256,
+  ATAK_FORWARDER = 257,
+  MAX = 511,
+}
